+++ crawler/core/config.py
@@ -1,4 +1,6 @@
 import yaml
+
+from loguru import logger
 from pathlib import Path
 
 
@@ -10,12 +12,12 @@ def config_read(name, msg="config"):
     try:
         config = yaml.safe_load(Path(name).read_text())
     except PermissionError:
-        print("ERROR: could not open config - please check file permissions")
+        logger.error("could not open config - please check file permissions")
         return None
     except yaml.YAMLError as error:
-        print("ERROR: %s" % error)
+        logger.error(error)
         return None
 
-    print("Successfully read %s from %s" % (msg, name))
+    logger.info("Successfully read %s from %s" % (msg, name))
 
     return config