+++ crawler/core/database.py
@@ -1,5 +1,7 @@
 import sys
 import sqlite3
+
+from loguru import logger
 from pathlib import Path
 
 
@@ -9,7 +11,7 @@ def database_connect(name):
         try:
             connection = sqlite3.connect(name)
         except sqlite3.OperationalError as error:
-            print("ERROR: %s" % error)
+            logger.error(error)
             return None
         return connection
     else:
@@ -23,7 +25,7 @@ def database_disconnect(connection):
 def database_initialize(name, prog_dirname):
     path = Path(name)
     if path.is_file():
-        print("WARNING: database %s already exists. Cowardly refusing action." % name)
+        logger.warning("database %s already exists. Cowardly refusing action." % name)
     else:
         create_statement_fqfn = prog_dirname + "/lib/initialize-image-catalog.sql"
         create_statement_file_path = Path(create_statement_fqfn)
@@ -32,21 +34,45 @@ def database_initialize(name, prog_dirname):
             create_statement = db_init_file.read()
             db_init_file.close()
         else:
-            raise SystemError("Template initialize-image-catalog.sql not found")
+            logger.error("Template initialize-image-catalog.sql not found")
+            raise SystemExit(1)
 
         try:
             connection = sqlite3.connect(name)
         except sqlite3.OperationalError as error:
-            print("ERROR: %s" % error)
+            logger.error(error)
         database_cursor = connection.cursor()
         try:
             database_cursor.execute(create_statement)
         except Exception as error:
-            print('ERROR: create table failed with the following error "%s"' % error)
+            logger.error('create table failed with the following error "%s"' % error)
 
         connection.close()
-        print("New database created under %s" % name)
+        logger.info("New database created under %s" % name)
 
+def db_get_last_checksum_fedora(connection, distribution):
+    try:
+        database_cursor = connection.cursor()
+        database_cursor.execute(
+            "SELECT checksum FROM image_catalog "
+            "WHERE distribution_name = '%s' "
+            "ORDER BY id DESC LIMIT 1" % distribution
+        )
+    except sqlite3.OperationalError as error:
+        logger.error(error)
+        raise SystemExit(1)
+
+    row = database_cursor.fetchone()
+
+    if row is None:
+        logger.debug("no previous entries found")
+        last_checksum = "sha256:none"
+    else:
+        last_checksum = row[0]
+
+    database_cursor.close()
+
+    return last_checksum
 
 def db_get_last_checksum(connection, distribution, release):
     try:
@@ -58,12 +84,13 @@ def db_get_last_checksum(connection, distribution, release):
             "ORDER BY id DESC LIMIT 1" % (distribution, release)
         )
     except sqlite3.OperationalError as error:
-        raise SystemError("SQLite error: %s" % error)
+        logger.error(error)
+        raise SystemExit(1)
 
     row = database_cursor.fetchone()
 
     if row is None:
-        # print("no previous entries found")
+        logger.debug("no previous entries found")
         last_checksum = "sha256:none"
     else:
         last_checksum = row[0]
@@ -87,7 +114,7 @@ def db_get_release_versions(connection, distribution, release, limit):
             "ORDER BY id DESC LIMIT %d" % (distribution, release, limit)
         )
     except sqlite3.OperationalError as error:
-        print("SQLite error: %s" % error)
+        logger.error(error)
         sys.exit(1)
     row = database_cursor.fetchone()
 
@@ -119,7 +146,8 @@ def read_version_from_catalog(connection, distribution, release, version):
             "ORDER BY ID" % (distribution, release, version)
         )
     except sqlite3.OperationalError as error:
-        raise SystemError("SQLite error: %s" % error)
+        logger.error(error)
+        raise SystemExit(1)
 
     image_catalog = {}
     image_catalog["versions"] = {}
@@ -154,7 +182,8 @@ def write_catalog_entry(connection, update):
         )
         connection.commit()
     except sqlite3.OperationalError as error:
-        raise SystemError("SQLite error: %s" % error)
+        logger.error(error)
+        raise SystemExit(1)
 
     database_cursor.close()
 
@@ -170,7 +199,8 @@ def update_catalog_entry(connection, update):
         )
         connection.commit()
     except sqlite3.OperationalError as error:
-        raise SystemError("SQLite error: %s" % error)
+        logger.error(error)
+        raise SystemExit(1)
 
     database_cursor.close()
 
@@ -186,7 +216,7 @@ def write_or_update_catalog_entry(connection, update):
     )
 
     if update["version"] in existing_entry["versions"]:
-        print("Updating version " + update["version"])
+        logger.info("Updating version " + update["version"])
         return update_catalog_entry(connection, update)
     else:
         return write_catalog_entry(connection, update)
@@ -195,16 +225,26 @@ def write_or_update_catalog_entry(connection, update):
 def read_release_from_catalog(connection, distribution, release, limit):
     try:
         database_cursor = connection.cursor()
-        database_cursor.execute(
-            "SELECT version,checksum,url,release_date "
-            "FROM (SELECT * FROM image_catalog "
-            "WHERE distribution_name = '%s' "
-            "AND distribution_release = '%s' "
-            "ORDER BY id DESC LIMIT %d) "
-            "ORDER BY ID" % (distribution, release, limit)
-        )
+        if release == "all":
+            database_cursor.execute(
+                "SELECT version,checksum,url,release_date,distribution_release "
+                "FROM (SELECT * FROM image_catalog "
+                "WHERE distribution_name = '%s' "
+                "ORDER BY id DESC LIMIT %d) "
+                "ORDER BY ID" % (distribution, limit)
+            )
+        else:
+            database_cursor.execute(
+                "SELECT version,checksum,url,release_date,distribution_release "
+                "FROM (SELECT * FROM image_catalog "
+                "WHERE distribution_name = '%s' "
+                "AND distribution_release = '%s' "
+                "ORDER BY id DESC LIMIT %d) "
+                "ORDER BY ID" % (distribution, release, limit)
+            )
     except sqlite3.OperationalError as error:
-        raise SystemError("SQLite error: %s" % error)
+        logger.error(error)
+        raise SystemExit(1)
 
     image_catalog = {}
     image_catalog["versions"] = {}
@@ -215,5 +255,6 @@ def read_release_from_catalog(connection, distribution, release, limit):
         image_catalog["versions"][version]["checksum"] = image[1]
         image_catalog["versions"][version]["url"] = image[2]
         image_catalog["versions"][version]["release_date"] = image[3]
+        image_catalog["versions"][version]["distribution_release"] = image[4]
 
     return image_catalog